[build-script] Don't rebuild CMake
This patch updates the CMake-building mechanism to avoid
re-bootstrapping CMake if we already bootstrapped one that is new
enough.

I've made it so that all paths through the function return the path to a
CMake so we can use the result of the function as the cmake path without
having to check.

The function will choose one of the following ways of getting CMake in
order of preference:
 - One we already built
 - The system CMake
 - Bootstrapping one from scratch

It prefers one we built over checking the system CMake because, if we
have already built a CMake previously, it's a good indication that
there either was no system CMake installed, or it wasn't new enough. We
shouldn't waste time checking it again if a previous run detected that
it wasn't good enough.

The system CMake is preferable to building one from scratch if we don't
need to though, so we determine if the system CMake is sufficient.

Finally, if one that we built either doesn't exist, or isn't new enough,
and the system either doesn't have a CMake, or a new enough CMake, build
one. It is built into the location that we are checking for caching, so
the next time we run build-script, it should hit the first case and
choose the already-built CMake instead of building it again.