[Diagnostics] Ambiguous argument diagnostic message
For ambiguous arguments, provide diagnostic with the set of
of options where applicable and improve message where the 
options are all the same.

Additionally consider options which are all functions with
the same number of parameters and provide an improved diagnostic
in these cases listing the types of the functions instead of 
not supporting these as similar argument types.