utils: enable android builds on Windows
Build and package the Android SDKs on Windows as part of the toolchain
distribution. This requires further fine tuning to improve the DevEx,
however, this ensures that we do not backslide on the ability to build
and package the SDK as a starting point.

Currently, we build for Android ARM64 only, but are likely to switch to
X64 to get test coverage.