IRGen: Fix assertion failure with typed throws
The mapTypeIntoContext() call in visitFullApplySite() was not
necessary.

The type in question is already fully substituted and should
no longer contain type parameters.

However, it can contain archetypes, which is what caused
mapTypeIntoContext() to assert.

Indeed, this case where the return type is generic but still
loadable wasn't covered by our test suite.

- Fixes https://github.com/swiftlang/swift/issues/80020.
- Fixes rdar://147051717.