[LLDB] Turn off resilience bypass for textual swiftinterface files
As the old comment indicates, it was not expected that build systems
would register binary swift modules produced from textual interfaces
with -add_ast_path, but we have found examples of the in the wild and
it leads to the most unexpected side effects. Fixing this is
straightforward, we can just check if a binary module was compiled
from a textual interface and then not bypass resilience.

rdar://145226754